Specify username and password for bootable targets
Before that, username and password of the current host user were copied
to the target. This created user 'ubuntu' when running from a live DVD
and was undesirable. Therefore prompts for username and password were
added. Username and password hash are saved in the config file.

In addition, function read_passphrase() now saves passphrases not only
per config file but also per auth method to the keyring.