Fix bugs: cannot use unformatted partition, no Ubuntu archive found
Before, trying to use an unformatted partition lead to 'Device is
mounted or has a holder or is unknown'. This was because blkid in
available_devs() apparently reports only devices that have content
metadata. Using ls instead solved the problem.

The host's Ubuntu archive URLs in /etc/apt/sources.list were assumed to
have a sub-subdomain which is not the case e.g. for live DVDs -- fixed.