Create a storage test script on every build/mount
Whenever run with -b or -m, stcomp.sh creates a script that can perform
read/write tests on all file systems, subvolumes and swap devices that
are part of the built/mounted configuration.

Each test script has the same path as the configuration file but has
extension '.test'. It consists of the configuration file and stcomp.fio
concatenated. Test scripts can be run in the host system, in chroot and
in the booted target system (they are copied to /usr/local/sbin).

Other improvements:
- handle config files containing spaces properly
- have uuid_to_devs() print _canonical_ names
- unmount in cleanup() only if TARGET points to a directory
- only warn if cache devices is non-removable or non-rotational
- abort if username missing on install
- write a header to the configuration file